chore: move tokenization tests to searchworker
